requirements.txt
Django==4.2.2
environs==9.5.0
python-dotenv==1.0.0
Pillow==9.5.0
django-admin-sortable2==2.1.8
